# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: EV Charging Station Booking API
  version: 1.0.0
  description: |
    REST API for the EV Charging Station Booking System.

    **Authentication** is session-based. Call `POST /api/auth/login` to receive a
    `JSESSIONID` cookie; include it in every subsequent request.

    **Roles**
    | Role   | Can do                                                         |
    |--------|----------------------------------------------------------------|
    | DRIVER | Register, log in, book connectors, manage own bookings.        |
    | ADMIN  | Everything a driver can do, plus manage stations, connectors, and all bookings. |
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://ev-charging-booking.herokuapp.com
    description: Production (Heroku)
...
paths:
...
  /api/stations:
    get:
      tags: [Stations]
      operationId: listStations
      summary: List all stations
      description: Returns every charging station with connector counts and today's booking count. No authentication required.
      responses:
        '200':
          description: Array of charging stations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StationResponse'
    post:
      tags: [Stations]
      operationId: createStation
      summary: Create a station
      description: Adds a new charging station. Requires **ADMIN** role.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StationRequest'
            example:
              name: "City Centre Hub"
              address: "1 High Street, London"
              latitude: 51.5074
              longitude: -0.1278
              description: "Underground car park, level 2."
      responses:
        '201':
          description: Station created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/stations/{id}:
    get:
      tags: [Stations]
      operationId: getStation
      summary: Get a station
      description: Returns a single charging station by ID. No authentication required.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: The charging station.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Stations]
      operationId: updateStation
      summary: Update a station
      description: Updates name, address, coordinates, or description of an existing station. Requires **ADMIN** role.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StationRequest'
      responses:
        '200':
          description: Updated station.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Stations]
      operationId: deleteStation
      summary: Delete a station
      description: Permanently deletes a station and all its connectors. Requires **ADMIN** role.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Station deleted — no content returned.
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Connectors ────────────────────────────────────────────────────────────
  /api/stations/{stationId}/connectors:
    get:
      tags: [Connectors]
      operationId: listConnectors
      summary: List connectors for a station
      description: Returns all connectors that belong to the specified station. No authentication required.
      parameters:
        - name: stationId
          in: path
          required: true
          description: ID of the parent station.
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Array of connectors.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConnectorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Connectors]
      operationId: createConnector
      summary: Add a connector to a station
      description: Creates a new connector for the given station. Requires **ADMIN** role.
      security:
        - cookieAuth: []
      parameters:
        - name: stationId
          in: path
          required: true
          description: ID of the parent station.
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectorRequest'
            example:
              connectorType: "CCS2"
              powerKw: 150
      responses:
        '201':
          description: Connector created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/connectors/{id}:
    get:
      tags: [Connectors]
      operationId: getConnector
      summary: Get a connector
      description: Returns a single connector by ID. No authentication required.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: The connector.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Connectors]
      operationId: updateConnector
      summary: Update a connector
      description: Updates the type and power rating of an existing connector. Requires **ADMIN** role.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectorRequest'
      responses:
        '200':
          description: Updated connector.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Connectors]
      operationId: deleteConnector
      summary: Delete a connector
      description: Permanently removes a connector. Requires **ADMIN** role.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Connector deleted — no content returned.
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/connectors/{id}/booked-times:
    get:
      tags: [Connectors]
      operationId: getBookedTimes
      summary: Get booked time windows
      description: |
        Returns all occupied time slots for a connector on a given calendar date.
        Used by the booking UI to visualise availability before creating a booking.
        No authentication required.
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: date
          in: query
          required: true
          description: Date to check, in ISO-8601 format (`YYYY-MM-DD`).
          schema:
            type: string
            format: date
            example: "2025-06-15"
      responses:
        '200':
          description: Array of occupied time windows for that date.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookedTimeResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Bookings ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Session cookie obtained after a successful `POST /api/auth/login`.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: Resource ID.
      schema:
        type: integer
        format: int64
        example: 1
  responses:
    BadRequest:
      description: Validation failed or request is malformed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 400
            error: "Bad Request"
            message: "must not be blank"
            timestamp: "2025-06-15T09:30:00"
...
    Forbidden:
      description: Authenticated but lacks the required role.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 403
            error: "Forbidden"
            message: "You do not have permission to perform this action."
            timestamp: "2025-06-15T09:30:00"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 404
            error: "Not Found"
            message: "Station not found with id: 99"
            timestamp: "2025-06-15T09:30:00"
...
  schemas:
...
    StationRequest:
      type: object
      required: [name, address, latitude, longitude]
      properties:
        name:
          type: string
          maxLength: 100
          example: "City Centre Hub"
        address:
          type: string
          maxLength: 200
          example: "1 High Street, London"
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
          example: 51.5074
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
          example: -0.1278
        description:
          type: string
          maxLength: 500
          nullable: true
          example: "Underground car park, level 2."
    ConnectorRequest:
      type: object
      required: [connectorType, powerKw]
      properties:
        connectorType:
          type: string
          example: "CCS2"
        powerKw:
          type: integer
          minimum: 1
          example: 150
...
    StationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "City Centre Hub"
        address:
          type: string
          example: "1 High Street, London"
        latitude:
          type: number
          format: double
          example: 51.5074
        longitude:
          type: number
          format: double
          example: -0.1278
        description:
          type: string
          nullable: true
          example: "Underground car park, level 2."
        connectorCount:
          type: integer
          example: 4
        todayBookingCount:
          type: integer
          example: 2
    ConnectorResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 3
        stationId:
          type: integer
          format: int64
          example: 1
        connectorType:
          type: string
          example: "CCS2"
        powerKw:
          type: integer
          example: 150
...
    BookedTimeResponse:
      type: object
      properties:
        startTime:
          type: string
          example: "10:00:00"
        endTime:
          type: string
          example: "11:30:00"
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "Not Found"
        message:
          type: string
          example: "Station not found with id: 99"
        timestamp:
          type: string
          format: date-time
          example: "2025-06-15T09:30:00"
...